Delete FwvTweak.esp
This is a very outdated version of my tweaks, left in for far too long. It is not installed via the latest fomod and only represents bloat. It is functionally identical to AdvancedModeRecall and completely replaced by it in future versions.

If needed for backwards compatibility on savegames (unlikely) then it can be found as version 1.2 on the Tweaks page on the nexus.