Delete FWVTimeCrackNoDisable
Note that this does NOT delete the Time Crack No Disable tweak. This is another bloat file left over from old versions of the mod which hasn't been overwritten yet because of a poor naming convention on my part. It is not installed by the fomod and is identical to the other tweak which does this and follows the correct naming convention.